backend/requirements.txt
fastapi==0.127.0
uvicorn==0.40.0
pytest==9.0.2
httpx==0.28.1
pytest-cov==7.0.0
black==25.12.0
isort==7.0.0
sqlmodel==0.0.27
reportlab==4.4.7
python-dotenv==1.2.1
apscheduler==3.11.1